Add LightnessSlider & apply()
Added LightnessSlider.pde and apply() method. Takes a selected range of 
lightnesses and adjusts them based on the slider position. It also 
adjusts pixels within a certain margin of the lightness range to avoid 
hard limits on adjustments made to the image.

+++ LiteRoom/Slider.pde
@@ -143,7 +143,7 @@ public abstract class Slider implements Interactable {
     }
   }
   
-  public float lightness(int r, int g, int b) {
+  public float lightness(float r, float g, float b) {
     float nR = r / 255;
     float nG = g / 255;
     float nB = b / 255;
